requirements.txt
discord~=1.0.1
requests~=2.28.1
python-dotenv~=0.14.0
sqlalchemy>=2.0.0
psycopg2
datetime~=5.0
feedparser~=6.0.10
Pillow~=10.0.1
regex==2023.10.3
pydantic~=1.10.2